requirements.txt
Flask==3.1.2
tensorflow==2.13.0
numpy==1.23.5
requests==2.32.3
gunicorn==21.2.0
Pillow==9.5.0
h5py==3.8.0
